Calculator with operations workish

+++ src/CalculationVisitor.java
@@ -1,25 +1,25 @@
 /**
  * Created by renju on 3/9/16.
  */
-public class CalculationVisitor  implements  Visitor{
+public class CalculationVisitor implements Visitor {
 
-    @Override
-    public Integer addVisit(Integer first, Integer second) {
-        return first + second;
-    }
+	@Override
+	public Double addVisit(Double first, Double second) {
+		return first + second;
+	}
 
-    @Override
-    public Integer subVisit(Integer first, Integer second) {
-        return first - second;
-    }
+	@Override
+	public Double subVisit(Double first, Double second) {
+		return first - second;
+	}
 
-    @Override
-    public Integer multVisit(Integer first, Integer second) {
-        return first * second;
-    }
+	@Override
+	public Double multVisit(Double first, Double second) {
+		return first * second;
+	}
 
-    @Override
-    public Integer divideVisit(Integer first, Integer second) {
-        return first / second;
-    }
+	@Override
+	public Double divideVisit(Double first, Double second) {
+		return first / second;
+	}
 }
