+++ src/CalculatorModel.java
@@ -1,29 +1,30 @@
 
 public class CalculatorModel implements Visitable {
-	private int calculatedValue;
+	private Double calculatedValue;
 	private CalculationVisitor visitor = new CalculationVisitor();
-	public void add(int first, int second) {
+
+	public void add(Double first, Double second) {
 		calculatedValue = visitor.addVisit(first, second);
 	}
-	
-	public void subtract(int first, int second) {
+
+	public void subtract(Double first, Double second) {
 		calculatedValue = visitor.subVisit(first, second);
 	}
 
-	public void multiply(int first, int second) {
+	public void multiply(Double first, Double second) {
 		calculatedValue = visitor.multVisit(first, second);
 	}
-	
-	public void divide(int first, int second) {
+
+	public void divide(Double first, Double second) {
 		calculatedValue = visitor.divideVisit(first, second);
 	}
-	
-	public int getCalculatedValue() {
+
+	public Double getCalculatedValue() {
 		return this.calculatedValue;
 	}
 
 	@Override
-	public Integer accept(Visitor visitor) {
+	public Double accept(Visitor visitor) {
 		return null;
 	}
 }
