+++ src/Visitable.java
@@ -2,5 +2,5 @@
  * Created by renju on 3/9/16.
  */
 public interface Visitable {
-    public Integer accept(Visitor visitor);
+	public Double accept(Visitor visitor);
 }
