+++ src/Visitor.java
@@ -2,8 +2,11 @@
  * Created by renju on 3/9/16.
  */
 public interface Visitor {
-    public Integer addVisit(Integer first, Integer second);
-    public Integer subVisit(Integer first, Integer second);
-    public Integer multVisit(Integer first, Integer second);
-    public Integer divideVisit(Integer first, Integer second);
+	public Double addVisit(Double first, Double second);
+
+	public Double subVisit(Double first, Double second);
+
+	public Double multVisit(Double first, Double second);
+
+	public Double divideVisit(Double first, Double second);
 }
